apps/ai_services/requirements.txt
# Core AI and Machine Learning Libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
joblib>=1.3.0

# Deep Learning and NLP Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
tokenizers>=0.13.0

# Vector Database and Search
faiss-cpu>=1.7.4
whoosh>=2.7.4

# Natural Language Processing
nltk>=3.8.0
spacy>=3.6.0
langdetect>=1.0.9
Sastrawi>=1.0.1  # Indonesian text processing

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
Pillow>=10.0.0
python-magic>=0.4.27

# Web Scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# Workflow automation
# n8n-python-sdk>=0.1.0  # Not available, using requests instead  # If available, otherwise use requests for N8N API

# Caching and Performance
redis>=4.6.0
python-memcached>=1.59

# Data Validation and Serialization
marshmallow>=3.20.0
cerberus>=1.3.4
validators>=0.20.0

# Async Processing
celery>=5.3.0
kombu>=5.3.0
billiard>=4.1.0

# Monitoring and Logging
psutil>=5.9.0
prometheus-client>=0.17.0
structlog>=23.1.0

# Configuration Management
python-decouple>=3.8
pyyaml>=6.0
toml>=0.10.2

# Testing Libraries
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
faker>=19.0.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.6

# Optional: GPU Support (uncomment if using CUDA)
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118
# faiss-gpu>=1.7.4

# Optional: Advanced NLP Models (uncomment if needed)
# spacy[transformers]>=3.6.0
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl

# Optional: Additional ML Libraries (uncomment if needed)
# xgboost>=1.7.0
# lightgbm>=4.0.0
# catboost>=1.2.0
# optuna>=3.3.0

# Optional: Time Series Analysis (uncomment if needed)
# statsmodels>=0.14.0
# prophet>=1.1.4
# pmdarima>=2.0.0

# Optional: Computer Vision (uncomment if needed)
# opencv-python>=4.8.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Optional: Advanced Text Processing (uncomment if needed)
# textblob>=0.17.1
# gensim>=4.3.0
# wordcloud>=1.9.0

# Optional: Database Connectors (uncomment if needed)
# psycopg2-binary>=2.9.0  # PostgreSQL
# pymongo>=4.4.0  # MongoDB
# elasticsearch>=8.8.0  # Elasticsearch

# Optional: Cloud Services (uncomment if needed)
# boto3>=1.28.0  # AWS
# google-cloud-storage>=2.10.0  # Google Cloud
# azure-storage-blob>=12.17.0  # Azure

# Optional: API Documentation (uncomment if needed)
# drf-spectacular>=0.26.0
# swagger-ui-bundle>=0.1.2